Moved common requests columns into a different file
