+++ SingularityUI/app/components/dashboard/DashboardPage.jsx
@@ -1,7 +1,7 @@
 import React, { Component, PropTypes } from 'react';
 import { connect } from 'react-redux';
 
-import RequestsTable from '../common/RequestsTable';
+import RequestsTable from '../requests/RequestsTable';
 
 
 class DashboardPage extends Component {
